Add additional benchmarks and unified benchmark script
